Struct binder::parcel::OwnedParcel

pub struct OwnedParcel { /* fields omitted */ }

The main type of Parcel. An OwnedParcel is (mostly) independant of the underlying BinderConnection : it owns its underlying data buffer.

Right now it still requires access to a BinderConnection though (which is also why BinderConnection is an Rc<RefCell>) in order to register handles to the connection's Handle HashMap. It is my hope to eventually break this dependency, perhaps by requiring the BinderConnection to be supplied as an argument to the read_*_binder functions.

Methods

impl OwnedParcel

fn new(conn: BinderConnection) -> OwnedParcel

fn len(&self) -> usize

unsafe fn set_data_len(&mut self, size: usize)

fn set_position(&mut self, pos: usize)

fn clear(&mut self)

fn write_i32(&mut self, val: i32)

fn write_u32(&mut self, val: u32)

fn write_pointer(&mut self, val: binder_uintptr_t)

fn write_interface_token(&mut self, interface: &str)

fn write_string16(&mut self, s: &str)

fn write_strong_binder(&mut self, handle: Option<Rc<RefCell<Handle>>>)

fn write_buf(&mut self, buf: &[u8])