Module binder::service_manager

Service Manager

On a default Android installation, the Service Manager is the first binder service you'll communicate with. It acts as a kind of directory for all other system-level services, and as a gateway, implementing some simple permission checking.

In most cases you'll only use the Service Manager to do one thing : get a connection to the Activity Manager service, which handles app-level services. However, it can be used for other purposes, given special privileges : you can get access to other system-level services such as the telephony services, or you can even register your own system-level service.

Structs

ServiceManager