Struct binder::BinderConnection

pub struct BinderConnection { /* fields omitted */ }

A connection to the Binder kernel interface.

The Binder connection doesn't actually allow to talk to other applications directly. Instead, you should first get an IBinder to the Service Manager by calling get_context_object(). You will then be able to access other services through the Service Manager.

Methods

impl BinderConnection

fn open() -> Result<BinderConnection>

Attempts to connection a connection to the Binder Kernel Interface.

Errors

This function will return :

WrongProtocolVersion if the kernel binder driver implements a different protocol version

Io if there is an error opening the connection to the driver

Nix if there is an error mmapping the Binder VM

fn get_service_manager(&mut self) -> Option<ServiceManager>

Gets a handle to the ServiceManager, or None if it wasn't registered as a context object yet. This can happen very early at boot, when the ServiceManager service hasn't been started yet.

Note that if your context object is not a ServiceManager (for instance, if you're not on Android), you should use get_context_object() instead.