Struct binder::Handle

pub struct Handle { /* fields omitted */ }

A connection to a Binder Service. Roughly equivalent to an sp in the libbinder framework.

In Binder parlance, there are two ways to be "connected" to a binder Service : We can have a Strong Pointer to a BbBinder instance, or we can have a Handle. In the first case, we say the Binder is "local", meaning it hasn't crossed process boundary. In the later case, we say the Binder is remote. When we send a Binder to another process, the kernel automatically transforms the pointer into a Handle. See https://github.com/torvalds/linux/blob/master/drivers/android/binder.c#L1594

For now, the only way to get one is to call BinderConnection.get_context_object

Methods

impl Handle

fn transact<'a, T: 'a + Parcel>(
&mut self,
code: u32,
data: &mut T,
flags: u32
) -> BinderResult<impl Parcel + 'a>