Function binder::sys::binder_version

pub unsafe fn binder_version(
    fd: c_int, 
    data: *mut binder_version
) -> Result<c_int>

Used to make sure that the version of binder we use is the same version the Kernel uses.

Note that there are two versions of the Binder protocol in the kernel tree. The version used is selected at kernel compile-time depending on the architecture used. 64-bit targets use version 8, while 32-bit use version 7.

In rust-binder, the version is chosen with the BINDER_IPC_32BIT cfg attribute.

TODO: Eventually, this will be exposed as a crate feature, or maybe even dynamically figured out (which would be an improvement over libbinder). But for now, we basically only allow version 7 (the 32-bit variant).